Add files via upload
A zipped folder containing files that we will use to analyze and extract any relevant data from. This  data is not mine, i retrieved the raw data from "The 20 Newsgroups data set". The 20 Newsgroups data set is a collection of approximately 20,000 newsgroup documents, divided across 20 different newsgroups. I only selected the files from the autos group. There are a total of 396 files (text files) in this folder. The source of the raw data is: http://qwone.com/~jason/20Newsgroups/.